more metadata cleanup
