+++ encodings/bytebool/src/compute.rs
