+++ encodings/dict/src/lib.rs
@@ -3,10 +3,10 @@
 //! Expose a [DictArray] which is zero-copy equivalent to Arrow's
 //! [DictionaryArray](https://docs.rs/arrow/latest/arrow/array/struct.DictionaryArray.html).
 pub use compress::*;
-pub use dict::*;
+pub use array::*;
 
 mod compress;
 mod compute;
-mod dict;
+mod array;
 mod stats;
 mod variants;
