+++ encodings/fastlanes/src/for/compress.rs
@@ -15,7 +15,8 @@ pub fn for_compress(array: &PrimitiveArray) -> VortexResult<Array> {
     let min = array
         .statistics()
         .compute(Stat::Min)
-        .ok_or_else(|| vortex_err!("Min stat not found"))?;
+        .ok_or_else(|| vortex_err!("Min stat not found"))?
+        .into_value();
 
     Ok(match_each_integer_ptype!(array.ptype(), |$T| {
         if shift == <$T>::PTYPE.bit_width() as u8 {
