+++ encodings/fastlanes/src/for/compute.rs
@@ -68,7 +68,7 @@ impl ScalarAtFn for FoRArray {
         let encoded =
             PrimitiveScalar::try_from(&encoded_scalar).vortex_expect("Invalid encoded scalar");
         let reference =
-            PrimitiveScalar::try_from(self.reference()).vortex_expect("Invalid reference scalar");
+            PrimitiveScalar::try_new(self.dtype(), self.reference()).vortex_expect("Invalid reference scalar");
 
         match_each_integer_ptype!(encoded.ptype(), |$P| {
             encoded.typed_value::<$P>().map(|v| (v << self.shift()).wrapping_add(reference.typed_value::<$P>().unwrap()))
@@ -105,6 +105,7 @@ fn search_sorted_typed<T>(
 where
     T: NativePType
         + for<'a> TryFrom<&'a Scalar, Error = VortexError>
+        + TryFrom<PValue, Error = VortexError>
         + Shr<u8, Output = T>
         + Shl<u8, Output = T>
         + WrappingSub
@@ -112,7 +113,7 @@ where
         + AddAssign
         + Into<PValue>,
 {
-    let min: T = array.reference().try_into()?;
+    let min: T = array.reference().as_pvalue()?.vortex_expect("Reference must be a valid pvalue").try_into()?;
     let primitive_value: T = value.cast(array.dtype())?.as_ref().try_into()?;
     // Make sure that smaller values are still smaller and not larger than (which they would be after wrapping_sub)
     if primitive_value < min {
