+++ encodings/fastlanes/src/for/mod.rs
@@ -10,7 +10,7 @@ use vortex::visitor::{AcceptArrayVisitor, ArrayVisitor};
 use vortex::{impl_encoding, Array, ArrayDType, ArrayTrait, Canonical, IntoCanonical};
 use vortex_dtype::{DType, PType};
 use vortex_error::{vortex_bail, vortex_panic, VortexExpect as _, VortexResult};
-use vortex_scalar::Scalar;
+use vortex_scalar::ScalarValue;
 
 mod compress;
 mod compute;
@@ -19,7 +19,7 @@ impl_encoding!("fastlanes.for", ids::FL_FOR, FoR);
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct FoRMetadata {
-    reference: Scalar,
+    reference: ScalarValue,
     shift: u8,
 }
 
@@ -30,17 +30,21 @@ impl Display for FoRMetadata {
 }
 
 impl FoRArray {
-    pub fn try_new(child: Array, reference: Scalar, shift: u8) -> VortexResult<Self> {
+    pub fn try_new(
+        child: Array,
+        reference: ScalarValue,
+        shift: u8,
+    ) -> VortexResult<Self> {
         if reference.is_null() {
-            vortex_bail!("Reference value cannot be null",);
+            vortex_bail!("Reference value cannot be null");
         }
-        let reference = reference.cast(
-            &reference
-                .dtype()
-                .with_nullability(child.dtype().nullability()),
-        )?;
+
+        if !reference.is_instance_of(child.dtype()) {
+            vortex_bail!("Reference value ({}) is not an instance of the child array dtype ({})", reference, child.dtype());
+        }
+
         Self::try_from_parts(
-            reference.dtype().clone(),
+            child.dtype().clone(),
             child.len(),
             FoRMetadata { reference, shift },
             [child].into(),
@@ -61,7 +65,7 @@ impl FoRArray {
     }
 
     #[inline]
-    pub fn reference(&self) -> &Scalar {
+    pub fn reference(&self) -> &ScalarValue {
         &self.metadata().reference
     }
